refactor(theme): consolidate design system, remove legacy theme
