+++ apps/flipcash/shared/theme/src/main/kotlin/com/flipcash/app/theme/FlipcashPreview.kt
@@ -9,18 +9,15 @@ import androidx.compose.runtime.CompositionLocalProvider
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.tooling.preview.PreviewWrapper
 import androidx.compose.ui.tooling.preview.PreviewWrapperProvider
 import com.getcode.animation.LocalSharedTransitionScope
 import com.flipcash.app.theme.internal.Flipcash2DesignSystem
-import com.flipcash.app.theme.internal.FlipcashLegacyDesignSystem
 import com.getcode.theme.CodeTheme
 
 @Deprecated("Use FlipcashThemeWrapper")
 @Composable
 fun FlipcashPreview(
     showBackground: Boolean = false,
-    useLegacyColors: Boolean = false,
     content: @Composable () -> Unit
 ) {
     val previewContent = @Composable {
@@ -39,11 +36,7 @@ fun FlipcashPreview(
         }
     }
 
-    if (useLegacyColors) {
-        FlipcashLegacyDesignSystem(previewContent)
-    } else {
-        Flipcash2DesignSystem(previewContent)
-    }
+    Flipcash2DesignSystem(previewContent)
 }
 
 class FlipcashThemeWrapper: PreviewWrapperProvider {
